Start extracting a base Message class.
